requirements.txt
discord.py==2.2.2
lupa==1.14.1